+++ app/errors/__init__.py
@@ -1,0 +1,3 @@
+from flask import Blueprint
+
+bp = Blueprint('errors', __name__)
