+++ app/feed/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+bp = Blueprint('feed', __name__)
+
+from app.feed import routes
