+++ app/myths/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+bp = Blueprint('myths', __name__)
+
+from app.myths import routes
