+++ app/questions/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+bp = Blueprint('questions', __name__)
+
+from app.questions import routes
