+++ app/scoring/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+bp = Blueprint('scoring', __name__)
+
+from app.scoring import routes
