+++ app/solutions/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+bp = Blueprint('solutions', __name__)
+
+from app.solutions import routes
