+++ app/values/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+bp = Blueprint('values', __name__)
+
+from app.values import routes
