+++ climatemind.py
@@ -1,6 +1,10 @@
-from knowledge_graph import app, db
-from knowledge_graph.models import Users, Scores, Sessions  # TODO Lrf, Zip, Iri
+from flask import current_app
+from app import create_app, db
+from app.models import Users, Scores, Sessions  # TODO Lrf, Zip, Iri
 
+
+app = create_app()
+app.app_context().push()
 
 @app.shell_context_processor
 def make_shell_context():
