--- knowledge_graph/make_network_test.py
@@ -1,6 +1,0 @@
-import pytest
-
-# import make_network
-
-# def test_answer():
-#    assert func(3) == 5
