--- knowledge_graph/test_loading_ontology.py
@@ -1,9 +1,0 @@
-import unittest
-
-from knowledge_graph.Mind import Mind
-
-
-class TestLoadingOntology(unittest.TestCase):
-    def test_valid_ontology_source(self):
-        onto = Mind()
-        self.assertIsNotNone(onto.get_ontology())
